% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/far_function.R
\name{fars_read}
\alias{fars_read}
\title{Function
Load and print data set
function that loads and print data set and check if it exist in R directory.}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{name of the file}
}
\value{
return the structure of file loaded as data frame
}
\description{
Function
Load and print data set
function that loads and print data set and check if it exist in R directory.
}
\examples{
\dontrun{
fars_read("C:\\\\data\\\\accident_2013.csv")
}
}
