% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/far_function.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Function
list of accident dataset for selected years
This function gives list of accident datasets with year.csv.bz2 structure}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{vector or list of selected dataset in month-year format}
}
\value{
returns to names accident datasets with the selected year in month-year format
}
\description{
Function
list of accident dataset for selected years
This function gives list of accident datasets with year.csv.bz2 structure
}
\examples{
\dontrun{
fars_read_years(2013, 2014, 2015)
fars_read_years("2013")
}
}
