% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/far_function.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Function
Summary statistics of accident each month of selected year
Gives summary statistics by year and month}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{selected years}
}
\value{
returns to a table with summary of accidents in each month of selected year
}
\description{
Function
Summary statistics of accident each month of selected year
Gives summary statistics by year and month
}
\examples{
\dontrun{
fars_summarize_years(c(2013, 2014, 2015))
fars_summarize_years("2013")
}
}
